Refactor product module with simplified form and listing
