+++ src/pages/Products.tsx
@@ -1,400 +1,16 @@
-import { useState } from "react";
-import { useForm } from "react-hook-form";
-import { z } from "zod";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useProducts } from "@/contexts/ProductContext";
-import { useSuppliers } from "@/contexts/SupplierContext";
-import { Product } from "@/types";
-
-import { PageHeader } from "@/components/common/PageHeader";
-import { Plus, Edit, Trash } from "lucide-react";
-import { Button } from "@/components/ui/button";
-import {
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
-import { Input } from "@/components/ui/input";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import { Textarea } from "@/components/ui/textarea";
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table";
-
-const productSchema = z.object({
-  name: z.string().min(1, "Nome é obrigatório"),
-  description: z.string().min(1, "Descrição é obrigatória"),
-  sku: z.string().min(1, "SKU é obrigatório"),
-  price: z.number().min(0.01, "Preço deve ser maior que zero"),
-  cost: z.number().min(0, "Custo não pode ser negativo"),
-  stockQuantity: z.number().int().min(0, "Quantidade não pode ser negativa"),
-  categoryId: z.string().min(1, "Categoria é obrigatória"),
-  supplierId: z.string().min(1, "Fornecedor é obrigatório"),
-  imageUrl: z.string().url("URL da imagem inválida").optional(),
-});
-
-type ProductFormValues = z.infer<typeof productSchema>;
+import { PageHeader } from '@/components/common/PageHeader';
+import ProductForm from '@/components/products/ProductForm';
+import ProductList from '@/components/products/ProductList';
 
 export default function Products() {
-  const { products, categories = [], createProduct, updateProduct, deleteProduct } = useProducts();
-  const { suppliers } = useSuppliers();
-  const [isDialogOpen, setIsDialogOpen] = useState(false);
-  const [editingProduct, setEditingProduct] = useState<Product | null>(null);
-  const [search, setSearch] = useState("");
-
-  const form = useForm<ProductFormValues>({
-    resolver: zodResolver(productSchema),
-    defaultValues: {
-      name: "",
-      description: "",
-      sku: "",
-      price: 0,
-      cost: 0,
-      stockQuantity: 0,
-      categoryId: "",
-      supplierId: "",
-      imageUrl: "",
-    },
-  });
-
-  const openNew = () => {
-    setEditingProduct(null);
-    form.reset({
-      name: "",
-      description: "",
-      sku: "",
-      price: 0,
-      cost: 0,
-      stockQuantity: 0,
-      categoryId: "",
-      supplierId: "",
-      imageUrl: "",
-    });
-    setIsDialogOpen(true);
-  };
-
-  const openEdit = (product: Product) => {
-    setEditingProduct(product);
-    form.reset({
-      name: product.name,
-      description: product.description,
-      sku: product.sku,
-      price: product.price,
-      cost: product.cost,
-      stockQuantity: product.stockQuantity,
-      categoryId: product.categoryId,
-      supplierId: product.supplierId,
-      imageUrl: product.imageUrl,
-    });
-    setIsDialogOpen(true);
-  };
-
-  const onSubmit = async (data: ProductFormValues) => {
-    try {
-      if (editingProduct) {
-        updateProduct({ ...editingProduct, ...data });
-      } else {
-        await createProduct({
-          ...data,
-          imageUrl: data.imageUrl ?? "",
-          isActive: true,
-        });
-      }
-
-      setIsDialogOpen(false);
-      setEditingProduct(null);
-      form.reset();
-      setSearch("");
-    } catch (error) {
-      console.error("Erro ao salvar produto:", error);
-    }
-  };
-
   return (
     <div className="space-y-6">
       <PageHeader
         title="Produtos"
-        description="Gerencie seu catálogo de produtos"
-        action={{
-          label: "Novo Produto",
-          onClick: openNew,
-          icon: <Plus className="h-4 w-4" />,
-        }}
-      >
-        <div className="pt-2">
-          <Input
-            placeholder="Buscar produto..."
-            value={search}
-            onChange={(e) => setSearch(e.target.value)}
-            className="max-w-sm"
-          />
-        </div>
-      </PageHeader>
-
-      <Table>
-        <TableHeader>
-          <TableRow>
-            <TableHead>Nome</TableHead>
-            <TableHead>Categoria</TableHead>
-            <TableHead>Quantidade</TableHead>
-            <TableHead>Preço</TableHead>
-            <TableHead>Fornecedor</TableHead>
-            <TableHead className="text-right">Ações</TableHead>
-          </TableRow>
-        </TableHeader>
-        <TableBody>
-          {products
-            .filter((p) =>
-              p.isActive &&
-              (p.name.toLowerCase().includes(search.toLowerCase()) ||
-                p.sku.toLowerCase().includes(search.toLowerCase()))
-            )
-            .map((p) => {
-              const category = categories.find((c) => c.id === p.categoryId);
-              const supplier = suppliers.find((s) => s.id === p.supplierId);
-              return (
-                <TableRow key={p.id}>
-                  <TableCell>{p.name}</TableCell>
-                  <TableCell>{category?.name || 'N/A'}</TableCell>
-                  <TableCell>{p.stockQuantity}</TableCell>
-                  <TableCell>
-                    {p.price.toLocaleString('pt-BR', {
-                      style: 'currency',
-                      currency: 'BRL',
-                    })}
-                  </TableCell>
-                  <TableCell>{supplier?.name || 'N/A'}</TableCell>
-                  <TableCell className="flex justify-end gap-2">
-                    <Button size="sm" variant="outline" onClick={() => openEdit(p)}>
-                      <Edit className="h-4 w-4" />
-                    </Button>
-                    <Button size="sm" variant="outline" onClick={() => deleteProduct(p.id)}>
-                      <Trash className="h-4 w-4" />
-                    </Button>
-                  </TableCell>
-                </TableRow>
-              );
-            })}
-        </TableBody>
-      </Table>
-
-      <Dialog open={isDialogOpen} onOpenChange={setIsDialogOpen}>
-        <DialogContent className="sm:max-w-[425px]">
-          <DialogHeader>
-            <DialogTitle>
-              {editingProduct ? "Editar Produto" : "Novo Produto"}
-            </DialogTitle>
-          </DialogHeader>
-          <Form {...form}>
-            <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
-              <FormField
-                control={form.control}
-                name="name"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Nome</FormLabel>
-                    <FormControl>
-                      <Input placeholder="Nome do produto" {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <FormField
-                control={form.control}
-                name="description"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Descrição</FormLabel>
-                    <FormControl>
-                      <Textarea
-                        placeholder="Descrição do produto"
-                        {...field}
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <FormField
-                control={form.control}
-                name="sku"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>SKU</FormLabel>
-                    <FormControl>
-                      <Input placeholder="SKU do produto" {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <div className="grid grid-cols-2 gap-4">
-                <FormField
-                  control={form.control}
-                  name="price"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Preço</FormLabel>
-                      <FormControl>
-                        <Input
-                          type="number"
-                          step="0.01"
-                          placeholder="0,00"
-                          {...field}
-                          onChange={(e) =>
-                            field.onChange(parseFloat(e.target.value))
-                          }
-                        />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-
-                <FormField
-                  control={form.control}
-                  name="cost"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Custo</FormLabel>
-                      <FormControl>
-                        <Input
-                          type="number"
-                          step="0.01"
-                          placeholder="0,00"
-                          {...field}
-                          onChange={(e) =>
-                            field.onChange(parseFloat(e.target.value))
-                          }
-                        />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </div>
-
-              <FormField
-                control={form.control}
-                name="stockQuantity"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Quantidade em Estoque</FormLabel>
-                    <FormControl>
-                      <Input
-                        type="number"
-                        placeholder="0"
-                        {...field}
-                        onChange={(e) =>
-                          field.onChange(parseInt(e.target.value))
-                        }
-                      />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <FormField
-                control={form.control}
-                name="categoryId"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Categoria</FormLabel>
-                    <Select
-                      onValueChange={field.onChange}
-                      defaultValue={field.value}
-                    >
-                      <FormControl>
-                        <SelectTrigger>
-                          <SelectValue placeholder="Selecione uma categoria" />
-                        </SelectTrigger>
-                      </FormControl>
-                      <SelectContent>
-                        {categories?.map((category) => (
-                          <SelectItem key={category.id} value={category.id}>
-                            {category.name}
-                          </SelectItem>
-                        ))}
-                      </SelectContent>
-                    </Select>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <FormField
-                control={form.control}
-                name="supplierId"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Fornecedor</FormLabel>
-                    <Select onValueChange={field.onChange} defaultValue={field.value}>
-                      <FormControl>
-                        <SelectTrigger>
-                          <SelectValue placeholder="Selecione um fornecedor" />
-                        </SelectTrigger>
-                      </FormControl>
-                      <SelectContent>
-                        {suppliers
-                          .filter((s) => s.isActive)
-                          .map((s) => (
-                            <SelectItem key={s.id} value={s.id}>
-                              {s.name}
-                            </SelectItem>
-                          ))}
-                      </SelectContent>
-                    </Select>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <FormField
-                control={form.control}
-                name="imageUrl"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>URL da Imagem</FormLabel>
-                    <FormControl>
-                      <Input placeholder="https://exemplo.com/imagem.jpg" {...field} />
-                    </FormControl>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-
-              <div className="flex justify-end gap-4">
-                <Button
-                  type="button"
-                  variant="outline"
-                  onClick={() => setIsDialogOpen(false)}
-                >
-                  Cancelar
-                </Button>
-                <Button type="submit">Salvar</Button>
-              </div>
-            </form>
-          </Form>
-        </DialogContent>
-      </Dialog>
+        description="Cadastre e visualize seus produtos"
+      />
+      <ProductForm />
+      <ProductList />
     </div>
   );
 }
